requirements.txt
flask
langchain
langchain-community
python-dotenv
pypdf
chromadb
beautifulsoup4
ollama
deep-translator

# openai==0.28.0
# pydantic==1.10.9
# yt-dlp
# pydub


# tiktoken
# langchain[docarray]

# SpeechRecognition
# openai-whisper
# gTTS

